requirements.txt
requests-aws4auth
opensearch-py
aws_lambda_powertools
aws-cdk-lib
boto3
sagemaker